Fixed bug in pmap_steal_memory
Fixed bug when calling PTE_CREATE, accidently passed in 'entry' variable
instead of kernel_pend which caused an incorrect mapping in the page
tables
Updated memset to set byte instead of uint32_t
Modifed ROUND_UP macro to NOT round up if already page-aligned

+++ include/string.h
@@ -8,7 +8,7 @@
 size_t strlen(const char *str);
 
 // Sets the first num bytes of a block of memory to the specified value
-void* memset(void *ptr, uint32_t value, size_t num);
+void* memset(void *ptr, uint8_t value, size_t num);
 
 // Copies num bytes from src to dst. Don't use this function if src and dst
 // overlap
